fix(Compression Node): Split utility tests into per-module test files
