+++ packages/nodes-base/nodes/Compression/test/node/BoundedUnzip.test.ts
@@ -1,13 +1,8 @@
 import * as fflate from 'fflate';
 
-import { boundedGunzip, boundedUnzip } from '../../utils';
+import { boundedUnzip } from '../../decompress/BoundedUnzip';
 
 type CompressionLevel = NonNullable<fflate.ZipOptions['level']>;
-
-function createGzipData(uncompressedSize: number): Buffer {
-	const data = new Uint8Array(uncompressedSize);
-	return Buffer.from(fflate.gzipSync(data));
-}
 
 function createZipData(
 	files: Record<string, number>,
@@ -34,42 +29,6 @@ function createZipWithUnsupportedCompression(): Buffer {
 
 	return compressed;
 }
-
-describe('boundedGunzip', () => {
-	it('should decompress data within the size limit', async () => {
-		const compressed = createGzipData(1024);
-		const result = await boundedGunzip(compressed, 2048);
-
-		expect(result).toBeInstanceOf(Buffer);
-		expect(result.length).toBe(1024);
-	});
-
-	it('should decompress data exactly at the size limit', async () => {
-		const compressed = createGzipData(1024);
-		const result = await boundedGunzip(compressed, 1024);
-
-		expect(result.length).toBe(1024);
-	});
-
-	it('should reject data exceeding the size limit', async () => {
-		const compressed = createGzipData(2048);
-
-		await expect(boundedGunzip(compressed, 1024)).rejects.toThrow(
-			'The decompressed output exceeds the maximum allowed size of 0 MB',
-		);
-	});
-
-	it('should handle empty gzip data', async () => {
-		const compressed = createGzipData(0);
-		const result = await boundedGunzip(compressed, 1024);
-
-		expect(result.length).toBe(0);
-	});
-
-	it('should reject empty gzip input', async () => {
-		await expect(boundedGunzip(Buffer.alloc(0), 1024)).rejects.toThrow('invalid gzip data');
-	});
-});
 
 describe('boundedUnzip', () => {
 	it('should decompress a zip with files within the size limit', async () => {
